Instrument mongo spring application
This commit instruments a spring boot application with mongo database
using the opentelemetry autoconfigured javaagent.

The traces exporter used is OTLP. This application is built using maven.

Command to build the app:
mvn clean install

The above command might report some errors while running the tests.
Those errors can be ignored as we're not interested in the resiliency of
the application. It builds the required jar file though.

Once the jar is ready, run docker compose up --build, jaeger ui is
exposed at port: 16686

While the mongo spring boot application listen at port 8080. The curl
commands to populate the database and fetch items are present in the
README.md file of this app.